+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Inject Patcher version into NVRAM for easier debugging
 - Add user-configurable ShowPicker
 - Add user-configurable Vaulting, enabled by default
+- Add user-configurable SIP and SecureBootModel
+- Fix USB Maps not working on "Minimal" SMBIOS
 
 ## 0.0.15
 - Add user-configurable OpenCore DEBUG builds